made some improvements

+++ libft/ft_strlcat.c
@@ -11,12 +11,13 @@
 /* ************************************************************************** */
 
 #include "libft.h"
+#include <sys/_types/_size_t.h>
 
 size_t	ft_strlcat(char *dest, char *src, size_t size)
 {
 	int		i;
 	size_t	j;
-	int		dest_len;
+	size_t	dest_len;
 
 	dest_len = ft_strlen(dest);
 	i = 0;
